Zork 4.3
I refactored the code so it would be divided amongst multiple classes and files so it will be easier to read and maintain.
I also changed how the rooms interact so the rooms are by specific rooms rather than by whatever room it happened to be placed next to in an array.